fix: viewディレクトリと該当ファイルの移動
