+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_training/start_screen.dart';
+import 'package:flutter_training/view/main_app.dart';
 
 void main() {
   runApp(
@@ -8,16 +8,4 @@ void main() {
       child: MainApp(),
     ),
   );
-}
-
-class MainApp extends StatelessWidget {
-  const MainApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const MaterialApp(
-      title: 'Weather App',
-      home: StartScreen(),
-    );
-  }
 }
